% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgov_ncts.R
\name{ctgov_ncts}
\alias{ctgov_ncts}
\title{Download trials by NCT}
\usage{
ctgov_ncts(nctids, batch_size = 500)
}
\arguments{
\item{nctids}{A list or a column of well-formed ClinicalTrials.gov
NCT Numbers, e.g. "NCT00942747". (A capitalized "NCT" followed
by eight numerals with no spaces or hyphens.)}

\item{batch_size}{The API will not function properly if you try to
pass a query that is too long, so this function breaks the NCT
Numbers provided into "batches" of 500 by default. The number
of NCT Numbers per query can be modified by changing this
variable.}
}
\value{
A nested list containing all the data provided by the
ClinicalTrials.gov API for the NCT Numbers provided.
}
\description{
Download a set of trial data from the ClinicalTrials.gov API by
their NCT Numbers
}
