% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgov_query.R
\name{ctgov_query}
\alias{ctgov_query}
\title{Download a set of trial data from the ClinicalTrials.gov API by
a search query}
\usage{
ctgov_query(query)
}
\arguments{
\item{query}{A character string containing a well-formed
ClinicalTrials.gov query in Essie expression syntax}
}
\value{
A nested list containing all the data provided by the
ClinicalTrials.gov API for the trials matching the search query
provided.
}
\description{
Download a set of trial data from the ClinicalTrials.gov API by
a search query
}
