% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgovtk-package.R
\docType{package}
\name{ctgovtk-package}
\alias{ctgovtk}
\alias{ctgovtk-package}
\title{ctgovtk: ClinicalTrials.gov Toolkit}
\description{
A toolkit providing various functions for interacting with ClinicalTrials.gov.
}
\details{
This package provides various functions for interacting
with ClinicalTrials.gov

ctgov_nct() returns a set of trial data from
ClinicalTrials.gov based on a search of NCT Numbers
}
\references{
Carlisle, BG. The grey literature, 2024.

#' @keywords internal
}
\author{
\strong{Maintainer}: Benjamin Gregory Carlisle \email{murph@bgcarlisle.com} (\href{https://orcid.org/0000-0001-8975-0649}{ORCID})

}
