% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_age_range.R
\name{extract_age_range}
\alias{extract_age_range}
\title{Extract age range}
\usage{
extract_age_range(ctgovdata)
}
\arguments{
\item{ctgovdata}{A nested list containing all the data provided by
the ClinicalTrials.gov API}
}
\value{
A data frame with one row per clinical trial and three
columns: nctid (a character string containing the NCT number
for the trial in question), min_age (a numeric value for the
minimum age for trial participants for the trial in question or
NA if none provided) and max_age (a numeric value for the
maximum age for trial participants for the trial in question or
NA if none provided)
}
\description{
Loops through all the entries in a set of trial data from the
ClinicalTrials.gov API, extracts the NCT number, minimum age and
maximum age for each and returns a data frame with those
data. Please note that this function does not download trials from
ClinicalTrials.gov; it parses trial data downloaded using the
ctgov_ncts() or ctgov_query() function and expects the ordered list
provided to be formatted in this manner.
}
