% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_basic_info.R
\name{extract_basic_info}
\alias{extract_basic_info}
\title{Extract basic trial info}
\usage{
extract_basic_info(ctgovdata)
}
\arguments{
\item{ctgovdata}{A nested list containing all the data provided by
the ClinicalTrials.gov API}
}
\value{
A data frame with one row per clinical trial and three
columns: nctid (a character string containing the NCT number for
the trial in question), brief_title (a character string
containing the brief title for the trial in question),
official_title (a character string containing the official title
for the trial in question), overall_status (a character string
containing the trial's overall status), phase (the phase of the
trial), enrol (enrolment count), enrol_type (ACTUAL or
ESTIMATED), min_age (minimum participant age as an integer or
NA), max_age (maximum participant age as an integer or NA), sex
(character string describing the sex of participants),
healthy_volunteers (Boolean, indicates whether healthy volunteers
are enrolled), study_type (INTERVENTIONAL or OBSERVATIONAL),
start_date (a character string containing an ISO-8601 formatted
date, YYYY-MM-DD or YYYY-MM corresponding to the trial start),
start_date_type (start date ACTUAL or ESTIMATED) pc_date (primary
completion date), pc_date_type (primary completion date ACTUAL or
ESTIMATED), fp_date (first posted date), fp_date_type (first
posted date ACTUAL or ESTIMATED)
}
\description{
Loops through all the entries in a set of trial data from the
ClinicalTrials.gov API, extracts basic trial info indexed by NCT
Number, one row per trial, and returns a data frame with those
data. Please note that this function does not download trials from
ClinicalTrials.gov; it parses trial data downloaded using the
ctgov_ncts() or ctgov_query() function and expects the ordered list
provided to be formatted in this manner.
}
