% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_docs.R
\name{extract_docs}
\alias{extract_docs}
\title{Extract attached study documents}
\usage{
extract_docs(ctgovdata)
}
\arguments{
\item{ctgovdata}{A nested list containing all the data provided by
the ClinicalTrials.gov API}
}
\value{
A data frame with one row per study document (there may be
multiple per NCT Number) and 11 columns: \code{nctid}, the NCT Number
for the trial in question, \code{typeAbbrev}, the abbreviation for the
type of study document, \code{hasProtocol}, a Boolean indicating
whether that document contains the study protocol, \code{hasSap}, a
Boolean indicating whether that document contains the statistical
analysis plan, \code{hasIcf}, a Boolean indicating whether that
document contains an informed consent form, \code{label}, a label for
the document, \code{date}, a date associated with the document,
\code{uploadDate}, the date that the document was uploaded to
ClinicalTrials.gov, \code{filename}, the filename associated with the
document in question, \code{size}, the number of bytes that the file
in question would take up on a hard disk and \code{url}, the hyperlink
to the download location of the document in question. In the case
that there is no file data for the set of documents provided, the
\code{url} column will not be generated and only 10 columns will be
returned.
}
\description{
Loops through all the entries in a set of trial data from the
ClinicalTrials.gov API, extracts study documents indexed by NCT
Number, one row per study document (meaning possibly more than one
row per trial), and returns a data frame with those data. Please
note that this function does not download trials from
ClinicalTrials.gov; it parses trial data downloaded using the
ctgov_ncts() or ctgov_query() function and expects the ordered list
provided to be formatted in this manner.
}
