% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_interventions.R
\name{extract_interventions}
\alias{extract_interventions}
\title{Extract interventions}
\usage{
extract_interventions(ctgovdata)
}
\arguments{
\item{ctgovdata}{A nested list containing all the data provided by
the ClinicalTrials.gov API}
}
\value{
A data frame with one row per intervention per trial and 6
columns: \code{nctid}, the NCT Number for the trial in question,
\code{type}, e.g. DRUG or BIOLOGICAL, \code{description} a character string
containing a description of the intervention, \code{armGroupLabels}, a
character string for a label for the group, \code{otherNames}, a
character string, for other names for the intervention
}
\description{
Extract interventions
}
