% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_outcome_measures.R
\name{extract_outcome_measures}
\alias{extract_outcome_measures}
\title{Extract outcome measures}
\usage{
extract_outcome_measures(ctgovdata)
}
\arguments{
\item{ctgovdata}{A nested list containing all the data provided by
the ClinicalTrials.gov API}
}
\value{
A data frame with one row per outcome measure per trial and
5 columns: \code{nctid}, the NCT Number for the trial in question,
\code{outcome_rank} (primary or secondary), \code{measure} (the name of the
outcome), \code{description} (a longer description of the outcome),
and \code{timeframe} (when the outcome is measured).
}
\description{
Loops through all the entries in a set of trial data from the
ClinicalTrials.gov API, extracts outcome measures indexed by NCT
Number, one row per outcome measure (meaning possibly more than one
row per trial), and returns a data frame with those data. Please
note that this function does not download trials from
ClinicalTrials.gov; it parses trial data downloaded using the
ctgov_ncts() or ctgov_query() function and expects the ordered list
provided to be formatted in this manner.
}
