% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_drug_name.R
\name{extract_drug_names}
\alias{extract_drug_names}
\title{Extract drug name}
\usage{
extract_drug_names(intervention)
}
\arguments{
\item{intervention}{A character string containing an unprocessed
intervention name (e.g. "Aspirin 150 mg" will return "Aspirin")
or NA in the case that there is none}
}
\value{
A character string or list containing the drug name(s)
found by the matching algorithm, if any
}
\description{
Extract drug name
}
